# Return NumPy array in from_iloc