--- a/tests/data/test_uniform_random_walker.py
+++ b/tests/data/test_uniform_random_walker.py
@@ -17,7 +17,7 @@
 import pytest
 import numpy as np
 from stellargraph.data.explorer import UniformRandomWalk
-from ..test_utils.graphs import create_test_graph
+from ..test_utils.graphs import create_test_graph, example_graph_random
 
 
 class TestUniformRandomWalk(object):
@@ -239,11 +239,11 @@ class TestUniformRandomWalk(object):
         assert np.array_equal(run_1, run_2)
 
     def test_benchmark_uniformrandomwalk(self, benchmark):
-
-        g = create_test_graph()
+        g = example_graph_random(n_nodes=100, n_edges=500)
         urw = UniformRandomWalk(g)
 
-        nodes = ["0"]  # this node has no edges including itself
+        nodes = np.arange(0, 50)
+        n = 2
         n = 5
         length = 5