Indexing Pandas types is slow, and indexing a `pandas.Index` is particularly slow, because it returns a whole `Index` object. This isn't generally useful for what we use `from_iloc` for, which is getting a sequence of node, edge or type IDs from some ilocs. The result is generally then used for things where having an `Index` isn't useful.

This makes the included microbenchmark of `from_iloc` (`test_benchmark_external_id_index_from_iloc`) significantly faster (times in microseconds):

| code | min | mean &pm; std | max | median &pm; IQR | 
|---|---|---|---|---|
| develop | 50.27 |  57.1 &pm; 9.7 | 224 | 52.9 &pm; 10 |
| this PR | 3.83 | 4.82 &pm; 3.1 | 64.6 | 4.25 &pm; 0.37 |

That is, a speed up of 12&times; going by the medians (ranging from 3.5-13&times; depending on which number one compares).

In #1267 991db681, random walks slowed down significantly compare to the previous commit (f2a96aab) due to more iloc conversions from storing edges as ilocs, see #1441.

In order to more accurately capture the effect of this change on those, the random walk benchmarks are increased to be slightly more realistic, with larger graphs and more head nodes. (#1444 does this change to the `BiasedRandomWalk` benchmarks, this does the rest.)

The `from_iloc` change thus has a significant impact on various benchmarks. The ones with the largest change are listed in the following table. The before/after columns are median times in microseconds &pm; IQR, and the speedup columns are comparing "after" with the specified before column.

| `test_benchmark_...` | before (pre-ilocs f2a96aab) | before (post-ilocs `develop`) | after | speedup vs. pre-ilocs | speedup vs. post-ilocs |
|---|---|---|---|---|--|
| `get_neighbours[False]`* | N/A | 16.1 &pm; 1.9 | 9.60 &pm; 1.3 | N/A | 1.7× |
| `biasedrandomwalk` | 7.79 &pm; 1.1 | 15.9 &pm; 2.1 | 8.12 &pm; 1.0 | 0.96× | 2.0× |
| `biasedweightedrandomwalk`^ | 83.7 &pm; 11 | 37.4 &pm; 5.7 | 29.2 &pm; 3.2 | 2.9× | 1.3× |
| `uniformrandommetapathwalk` | 235 &pm; 26 | 112 &pm; 11 | 39.4 &pm; 4.6 | 6.0× | 2.8× |
| `uniformrandomwalk` | 10.6 &pm; 1.3 | 31.8 &pm; 1.3 | 14.0 &pm; 1.4 | 0.76× | 2.3× |

\* `test_benchmark_get_neighbours[use_ilocs=False]` didn't exist before ilocs
^ `BiasedRandomWalk(..., weighted=True)` has independent optimisations applied in #1444, included here.

Some other benchmarks have smaller improvements (speedups up to ~1.1&times;), such as the GraphSAGE and HinSAGE random walkers.

In summary, this is a 1.3&times;-2.8&times; speedups on many random walk benchmarks, bringing the benchmarks with slowdowns in #1441 to be about what they were before ilocs were used for edges in #1267 991db681. It likely reduces memory use too, due to not needing the metadata of the `Index` type.

For a real world dataset, continuing the Cora table from #1444:

| code | unweighted (s) | weighted (s) |
|---|---|---|
| f2a96aab (pre-iloc)  | 1.36 ± 0.0233 | 11.7 ± 0.0724 |
| 991db681 (iloc) | 2.02 ± 0.0173 | 13 ± 0.427 |
| #1444 | 2 ± 0.101 | 5.64 ± 0.179  |
| This PR | 1.46 ± 0.0376 | 5.46 ± 0.0562 |

See: #1441